refactor connection point funcs into separate dir
