+++ lib/types/srj-types.ts
@@ -78,37 +78,9 @@ export interface SimplifiedPcbTrace {
 
 export type SimplifiedPcbTraces = Array<SimplifiedPcbTrace>
 
-// Type guards and helpers for ConnectionPoint types
-export function isMultiLayerConnectionPoint(
-  point: ConnectionPoint,
-): point is MultiLayerConnectionPoint {
-  return "layers" in point && Array.isArray((point as any).layers)
-}
-
-export function isSingleLayerConnectionPoint(
-  point: ConnectionPoint,
-): point is SingleLayerConnectionPoint {
-  return "layer" in point && typeof (point as any).layer === "string"
-}
-
-/**
- * Gets the primary layer from a connection point.
- * For MultiLayerConnectionPoint, returns the first layer as default.
- */
-export function getConnectionPointLayer(point: ConnectionPoint): string {
-  if (isMultiLayerConnectionPoint(point)) {
-    return point.layers[0]
-  }
-  return point.layer
-}
-
-/**
- * Gets all layers from a connection point.
- * For ConnectionPoint, returns an array with the single layer.
- */
-export function getConnectionPointLayers(point: ConnectionPoint): string[] {
-  if (isMultiLayerConnectionPoint(point)) {
-    return point.layers
-  }
-  return [point.layer]
-}
+export {
+  isMultiLayerConnectionPoint,
+  isSingleLayerConnectionPoint,
+  getConnectionPointLayer,
+  getConnectionPointLayers,
+} from "../utils/connection-point-utils"
